+++ tests/test_mda.py
@@ -126,10 +126,7 @@ class TestMDA:
         # universe itself, which isn't great
         u = MNUniverse
         bpy.context.scene.frame_set(0)
-        u.object.mn_universe_selections.add()
-        sel = u.object.mn_universe_selections[0]
-        sel.name = "custom_sel_1"
-        sel.selection_str = "around 3.5 protein"
+        sel = u.add_selection(name="custom_sel_1", selection_str="around 3.5 protein")
         bpy.context.scene.frame_set(5)
         sel_1 = u.named_attribute("custom_sel_1")
         bpy.context.scene.frame_set(50)
